requirements.txt
Pillow 
google-genai
tqdm